+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### X.X.X (comming soon)
+### 8.7.2 (24 October 2019)
 
 Bugfixes:
   - GraphQL replication sometimes not pushes when a big amount of documents has been pulled before